Delete Cannon_140mm_x2.ini
Not sure why this ever existed. Apparently it was meant for the Overlord Tank, created under the impression that its cannons were 140mm. The cannons are already covered in entries in 120mm and 130mm (currently being revised and combined), so I am moving it there and fusing it with them.