Delete Grom.txt
Merged this into the Igla. Not sure why there was an entry for the Grom anyway, which is a Polish variant of the 9K38 Igla, the weapon that would be in common Soviet use, especially since (aligned with RotR) to implement "Igla Conscripts", and thus the Igla itself, was the more likely route.